+++ HISTORY.md
@@ -2,6 +2,18 @@
 
 -----
 
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## Unreleased
+
+***Added:***
+
+- Improve command execution such that the currently defined environment variable filters are always reflected
+- Commands now run within a shell so that expansion semantics are supported (such as `$ENV_VAR`)
+- Bump the minimum supported version of Hatch
+
 ## 0.1 - 2021-12-29
 
 This is the initial public release.